requirements.txt
setuptools
numpy
matplotlib
scipy
h5py
lsl
wxpython <= 4.1.1
importlib-metadata; python_version < '3.8'
healpy < 1.14.0; python_version <= '2.7'
bottleneck; python_version > '2.7'
